See http://sandbox.napa-ibs.net for working site
Need the IBS logo to be the clickable home page link
- Done

For the top menu items, change the hover event to only “grow” the icon
(like done in for the carousel). The menu will only expand on click of the
icon (when no href is provided)
- Enabled initial click, then switches to hovers
- Enabled the border only when active

I reworked the top menu layout a bit to provide the Fleet Manager set
(followed basecamp patterns as I understand them). Anyway in some
browsers the bottom boarder is 1px (or so) up from the rainbow bar.

Stack the IBS Member Login and the search icon buttons on separate
lines (still right justified) and removed the separator.
- Done

Would like the ability to have either a border (rounded corners) or a
background color behind the Welcome Mat element (see sandbox)
- Added a css rule for this

Is there an option to center the galleria thumbnails.
- Done

Update the border color so that buttons are more distinctive
-Done

Also how do I fix the No theme message. Obviously the relative path to a css is
broken, but I couldn’t figure out which are where to update it.
- I think it was in /js/vendor/galleria.ibs.js (which is just a renamed classic theme)

We need the gears from the mockups and prototype on the left and
right page borders. I have voiced concerns about this, but still lost the
battle. Hopefully you have a better why to do it than the hack I used for
the prototype.
- I need the original image to do this.  Can't get it from the pdfs